FlaskServer/requirements.txt
Flask
pandas
numpy
sklearn
umap-learn
scipy
openai
python-dotenv
